requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
fpdf2>=2.7.0
pytest>=7.4.0
